Удалил скрипт start.sh и перенастроил Dockerfile для запуска Uvicorn напрямую. Реорганизовал структуру приложения, добавив маршруты для категорий и продуктов, а также вынес функции работы с Redis в отдельный файл. Обновил схемы Pydantic для проверки значений цены товара.
